+++ src/activity/components/createDesafioClasificacion/conceptList/ConceptList.module.css
@@ -1,0 +1,169 @@
+.content {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  padding: 0 1.5rem 1.5rem 1.5rem;
+}
+
+.conceptsSection {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+}
+
+.conceptsHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 1rem;
+}
+
+.conceptsTitle {
+  font-size: 1rem;
+  font-weight: 600;
+  color: #2d3748;
+  margin: 0;
+}
+
+.addConceptButton {
+  padding: 0.5rem 1rem;
+  font-size: 0.875rem;
+}
+
+.addConceptForm {
+  display: flex;
+  gap: 0.5rem;
+  align-items: flex-start;
+  margin-bottom: 1rem;
+  padding: 1rem;
+  background: #f7fafc;
+  border-radius: 8px;
+  border: 2px dashed #cbd5e0;
+}
+
+.conceptInput {
+  flex: 1;
+}
+
+.addConceptActions {
+  display: flex;
+  gap: 0.25rem;
+  flex-shrink: 0;
+}
+
+.conceptsList {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  min-height: 100px;
+}
+
+.conceptItem {
+  background: #f7fafc;
+  border: 1px solid #e2e8f0;
+  border-radius: 8px;
+  padding: 0.75rem;
+  transition: all 0.2s ease;
+}
+
+.conceptItem:hover {
+  background: #edf2f7;
+  border-color: #cbd5e0;
+}
+
+.editConceptForm {
+  display: flex;
+  gap: 0.5rem;
+  align-items: center;
+}
+
+.conceptContent {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.conceptName {
+  flex: 1;
+  font-size: 0.875rem;
+  color: #2d3748;
+  word-break: break-word;
+}
+
+.conceptActions {
+  display: flex;
+  gap: 0.25rem;
+  opacity: 0;
+  transition: opacity 0.2s ease;
+}
+
+.conceptItem:hover .conceptActions {
+  opacity: 1;
+}
+
+.conceptActionButton {
+  padding: 0.25rem;
+  min-width: auto;
+  width: 28px;
+  height: 28px;
+  font-size: 0.75rem;
+}
+
+.conceptActionButton.deleteButton:hover {
+  background-color: #fed7d7;
+  color: #c53030;
+}
+
+.emptyConcepts {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 2rem;
+  text-align: center;
+  color: #718096;
+  border: 2px dashed #cbd5e0;
+  border-radius: 8px;
+  background: #f7fafc;
+}
+
+.emptyConceptsText {
+  font-size: 0.875rem;
+  font-weight: 500;
+  margin: 0 0 0.25rem 0;
+}
+
+.emptyConceptsHint {
+  font-size: 0.75rem;
+  margin: 0;
+  opacity: 0.8;
+}
+
+/* Responsive Design */
+@media (max-width: 768px) {
+  .conceptsHeader {
+    flex-direction: column;
+    gap: 0.75rem;
+    align-items: stretch;
+  }
+
+  .addConceptButton {
+    width: 100%;
+  }
+
+  .addConceptForm {
+    flex-direction: column;
+    gap: 0.75rem;
+  }
+
+  .addConceptActions {
+    justify-content: center;
+  }
+
+  .conceptActions {
+    opacity: 1;
+  }
+}
